chore: Rename the header of adrc impl
