+++ rmcs_ws/src/rmcs_core/src/controller/adrc/adrc_controller.cpp
@@ -2,14 +2,12 @@
 #include <cmath>
 #include <limits>
 
-#include <rclcpp/logging.hpp>
 #include <rclcpp/node.hpp>
-#include <rclcpp/rclcpp.hpp>
 #include <rmcs_executor/component.hpp>
 
-#include "ESO.hpp"
-#include "NLESF.hpp"
-#include "TD.hpp"
+#include "eso.hpp"
+#include "nlesf.hpp"
+#include "td.hpp"
 
 namespace rmcs_core::controller::adrc {
